+++ sesman/sesexec/Makefile.am
@@ -19,6 +19,12 @@ pkglibexec_PROGRAMS = \
 xrdp_sesexec_SOURCES = \
   sesexec.c \
   sesexec.h \
+  session_base.c \
+  session_base.h \
+  session_x11.c \
+  session_x11.h \
+  session_parameters.c \
+  session_parameters.h \
   session.c \
   session.h \
   ccp_server.c \
